moved recursive reply adding, deleting, replacing functions to util folder
